refactor: my-book-detail を my-book に移動
