+++ app/build.gradle.kts
@@ -136,7 +136,6 @@ dependencies {
   implementation(project(":core:supabase"))
   implementation(project(":core:ui"))
   implementation(project(":feature:my-book"))
-  implementation(project(":feature:my-book-detail"))
   implementation(project(":feature:search-book"))
   implementation(project(":feature:send-otp"))
   implementation(project(":feature:setting"))
