+++ core/domain/src/main/java/app/kaito_dogi/mybrary/core/domain/model/MyBook.kt
@@ -18,4 +18,4 @@ data class MyBook(
   val isFavorite: Boolean,
   val isPublic: Boolean,
   val isArchived: Boolean,
-) : java.io.Serializable
+)
