+++ core/ui/src/main/java/app/kaito_dogi/mybrary/core/ui/navigation/MybraryRoute.kt
@@ -1,17 +1,11 @@
 package app.kaito_dogi.mybrary.core.ui.navigation
 
-import app.kaito_dogi.mybrary.core.domain.model.MyBook
 import kotlinx.serialization.Serializable
 
 @Deprecated("")
 sealed interface MybraryRoute {
   @Serializable
   data object MyBookList : MybraryRoute
-
-  @Serializable
-  data class MyBookDetail(
-    val myBook: MyBook,
-  ) : MybraryRoute
 
   @Serializable
   data object SearchBook : MybraryRoute
