+++ feature/my-book/src/main/java/app/kaito_dogi/mybrary/feature/mybook/destination/mybooklist/MyBookListNavigation.kt
@@ -16,4 +16,4 @@ fun NavGraphBuilder.myBookListScreen(
   )
 }
 
-fun NavHostController.navigateToMyBookListScreen() = this.navigate(MyBookRoute.MyBookList)
+fun NavHostController.navigateToMyBookListScreen() = this.navigate(route = MyBookRoute.MyBookList)
