+++ feature/my-book/src/main/java/app/kaito_dogi/mybrary/feature/mybook/destination/mybooklist/MyBookListScreen.kt
@@ -46,7 +46,6 @@ internal fun MyBookListScreen(
         )
       }
     },
-    contentWindowInsets = WindowInsets.systemBars.add(WindowInsets(bottom = 80.dp)),
   ) { innerPadding ->
     LazyVerticalGrid(
       columns = GridCells.Fixed(uiState.numberOfColumns),
