+++ settings.gradle.kts
@@ -26,7 +26,6 @@ include(":core:domain")
 include(":core:supabase")
 include(":core:ui")
 include(":feature:my-book")
-include(":feature:my-book-detail")
 include(":feature:search-book")
 include(":feature:send-otp")
 include(":feature:setting")
